tools/audio_analyzer/requirements.txt
flask==2.3.2
requests==2.31.0
numpy==1.26.4
soundfile==0.12.1
webrtcvad==2.0.10
vosk==0.3.47
torch>=2.2.0
openai-whisper==20230314
whisperx==0.3.7
librosa==0.10.2
scipy==1.11.3
typing-extensions
